New Release --------- Added abstraction and interface

+++ src/Assignment/TotalPayment.java
@@ -1,4 +1,4 @@
 package Assignment;
 public interface TotalPayment {
-    public double calTotalPayment();
+    public int calTotalPayment();
 }
